refactor: move tests to one folder
