+++ src/tests/mocks/TestSetup.json
@@ -1,0 +1,55 @@
+{
+  "_format": "hh-sol-artifact-1",
+  "contractName": "TestSetup",
+  "sourceName": "contracts/TestSetup.sol",
+  "abi": [
+    {
+      "inputs": [],
+      "stateMutability": "payable",
+      "type": "constructor"
+    },
+    {
+      "anonymous": false,
+      "inputs": [
+        {
+          "indexed": false,
+          "internalType": "uint256",
+          "name": "amount",
+          "type": "uint256"
+        },
+        {
+          "indexed": false,
+          "internalType": "uint256",
+          "name": "when",
+          "type": "uint256"
+        }
+      ],
+      "name": "Withdrawal",
+      "type": "event"
+    },
+    {
+      "inputs": [],
+      "name": "owner",
+      "outputs": [
+        {
+          "internalType": "address payable",
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "stateMutability": "view",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "withdraw",
+      "outputs": [],
+      "stateMutability": "nonpayable",
+      "type": "function"
+    }
+  ],
+  "bytecode": "0x6080604052336000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506101fd806100536000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c80633ccfd60b1461003b5780638da5cb5b14610045575b600080fd5b610043610063565b005b61004d610105565b60405161005a919061016a565b60405180910390f35b7fbf2ed60bd5b5965d685680c01195c9514e4382e28e3a5a2d2d5244bf59411b93474260405161009492919061019e565b60405180910390a160008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc479081150290604051600060405180830381858888f19350505050158015610102573d6000803e3d6000fd5b50565b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b600061015482610129565b9050919050565b61016481610149565b82525050565b600060208201905061017f600083018461015b565b92915050565b6000819050919050565b61019881610185565b82525050565b60006040820190506101b3600083018561018f565b6101c0602083018461018f565b939250505056fea26469706673582212204dadbb4c4d805fd0cb115b6c1d8897af72cecb2e4f9a52e593c76b8524c46e4964736f6c63430008110033",
+  "deployedBytecode": "0x608060405234801561001057600080fd5b50600436106100365760003560e01c80633ccfd60b1461003b5780638da5cb5b14610045575b600080fd5b610043610063565b005b61004d610105565b60405161005a919061016a565b60405180910390f35b7fbf2ed60bd5b5965d685680c01195c9514e4382e28e3a5a2d2d5244bf59411b93474260405161009492919061019e565b60405180910390a160008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc479081150290604051600060405180830381858888f19350505050158015610102573d6000803e3d6000fd5b50565b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b600061015482610129565b9050919050565b61016481610149565b82525050565b600060208201905061017f600083018461015b565b92915050565b6000819050919050565b61019881610185565b82525050565b60006040820190506101b3600083018561018f565b6101c0602083018461018f565b939250505056fea26469706673582212204dadbb4c4d805fd0cb115b6c1d8897af72cecb2e4f9a52e593c76b8524c46e4964736f6c63430008110033",
+  "linkReferences": {},
+  "deployedLinkReferences": {}
+}
